refactor to toggle like and dislikes

+++ app/Http/Controllers/Api/v1/QuestionFavoriteController.php
@@ -12,7 +12,7 @@ class QuestionFavoriteController extends Controller
         $favorites = $question->favorites()->where('user_id', auth()->id())->get();
 
         if($favorites->count() > 0){
-            $favorites->delete();
+            $question->favorites()->where('user_id', auth()->id())->delete();
         }
         
         if($favorites->count() == 0){
